Removed unecessary component, all related to employee table is now in employee table component
