+++ src/components/User/EmployeeTable.tsx
@@ -1,8 +1,29 @@
-import { Flex } from "@chakra-ui/react";
-import Employees from "./Employees";
+import {
+  Box,
+  Flex,
+  Table,
+  Tbody,
+  Text,
+  Th,
+  Thead,
+  Tr,
+  useColorModeValue,
+} from "@chakra-ui/react";
+import EmployeeTableRow from "./EmployeeTableRow";
+
+type EmployeeInfo = {
+  email: string;
+  name: string;
+  role: string;
+  active: boolean;
+  date: string;
+};
 
 const EmployeeTable = () => {
-  const tablesTableData = [
+  const textColor = useColorModeValue("gray.700", "white");
+  const captions = ["Employee", "Function", "Status", "Employed", ""];
+
+  const employeesData = [
     {
       name: "Marko Markovic",
       email: "marko@example.com",
@@ -63,11 +84,46 @@ const EmployeeTable = () => {
 
   return (
     <Flex direction="column" pt={{ base: "30px", md: "0px" }}>
-      <Employees
-        title={"Employees Table"}
-        captions={["Employee", "Function", "Status", "Employed", ""]}
-        data={tablesTableData}
-      />
+      <Box overflowX={{ sm: "scroll", xl: "hidden" }}>
+        <Box p="6px 0px 22px 0px">
+          <Text fontSize="xl" color={textColor} fontWeight="bold">
+            {"Employees Table"}
+          </Text>
+        </Box>
+        <Box>
+          <Table variant="simple" color={textColor}>
+            <Thead>
+              <Tr my=".8rem" pl="0px" color="gray.400">
+                {captions.map((caption: string, index: number) => {
+                  return (
+                    <Th
+                      color="gray.400"
+                      key={index}
+                      ps={index === 0 ? "0px" : undefined}
+                    >
+                      {caption}
+                    </Th>
+                  );
+                })}
+              </Tr>
+            </Thead>
+            <Tbody>
+              {employeesData.map((row: EmployeeInfo) => {
+                return (
+                  <EmployeeTableRow
+                    key={`${row.email}-${row.name}`}
+                    name={row.name}
+                    email={row.email}
+                    role={row.role}
+                    active={row.active}
+                    date={row.date}
+                  />
+                );
+              })}
+            </Tbody>
+          </Table>
+        </Box>
+      </Box>
     </Flex>
   );
 };
